+++ tests/testthat.R
@@ -1,3 +1,0 @@
-library(testthat)
-library(maritimeinformatics)
-test_check("maritimeinformatics")